refactoring the form component
